+++ classes/fruit.class.js
