+++ game/snake.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title>Snake Game</title>
+        <script type="module" src="game.js"></script>
+        <link rel="stylesheet" href="style.css">
+    </head>
+    <body>
+        <div class="container">
+            <div class="game-btn">
+                <button id="start">Start</button>
+                <button id="finish">Finish</button>
+            </div>
+            <canvas id="canvas" width="600px" height="600px"></canvas>
+        </div>
+    </body>
+</html>
