+++ game/style.css
@@ -1,0 +1,55 @@
+* {
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+    outline: none;
+}
+body {
+    width: 80%;
+    height: 80%;
+
+    display: flex;
+    flex-direction: column;
+
+    background: url(../img/background.jpg) no-repeat;
+}
+button {
+    border: none;
+    cursor: pointer;
+    background-color: #054a29;
+    border-radius: 50px;
+
+    width: 200px;
+    height: 50px;
+
+    color: white;
+    font-size: 25px;
+    font-weight: 500;
+}
+
+.container {
+    display: flex;
+    width: 100vw;
+    height: 100vh;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 30px;
+}
+
+.game-btn {
+    display: flex;
+    gap: 30px;
+}
+
+#canvas {
+    background-color: black;
+    z-index: -1;
+    display: none;
+    border: 5px solid white;
+}
+
+#finish {
+    display: none;
+    background-color: #fe4a49;
+}
