--- snake.html
@@ -1,77 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Snake Game</title>
-        <style>
-            * {
-                margin: 0;
-                padding: 0;
-                font-family: sans-serif;
-                outline: none;
-            }
-            body {
-                width: 80%;
-                height: 80%;
-
-                display: flex;
-                flex-direction: column;
-
-                background: url(./background.jpg) no-repeat;
-            }
-            button {
-                border: none;
-                cursor: pointer;
-                background-color: #054A29;
-                border-radius: 50px;
-
-                width: 200px;
-                height: 50px;
-
-                color: white;
-                font-size: 25px;
-                font-weight: 500;
-            }
-            
-            .container {
-                display: flex;
-                width: 100vw;
-                height: 100vh;
-                flex-direction: column;
-                align-items: center;
-                justify-content: center;
-                gap: 30px;
-            }
-            
-            .game-btn {
-                display: flex;    
-                gap: 30px;
-            }
-            
-            #canvas {
-                background-color: black;
-                z-index: -1;
-                display: none;
-                border: 5px solid white;
-            }
-            
-            #finish {
-                display: none;
-                background-color: #FE4A49;
-            }
-        </style>
-        <script type="module" src="game.js"></script>
-    </head>
-    <body>
-        <div class="container">
-            <div class="game-btn">
-                <button id="start">Start</button>
-                <button id="finish">Finish</button>
-                
-            </div>
-            <canvas id="canvas" width="600px" height="600px"></canvas>
-        </div>
-        <!-- <script type="module" src="game.js"></script> -->
-    </body>
-</html>
